admin: Add new Astakos states
The user will be considered to be in ONE of the following states:

Active: The user is activated and moderated.
Inactive: The user is moderated but may be rejected or inactive.
Pending moderation: The email of the user is verified but the user is
not moderated.
Pending verification: The user's mail is not verified.

This state is reflected in the user list as well as in the <tr> of the
html template.