admin: Create generic class for model actions
Add AdminAction class, which can be used by any model-view to create
actions for this model. The AdminAction class should be generic enough
and by default provides the following attributes:

    Attributes:
        op:             The action
        name:           The name of the action
        description:    A short text that describes an action
        target:         The target of the action
        severity:       The negative impact of the action.
                        Accepted values: trivial, big, irreversible
        allowed_groups: The groups that are allowed to author this
                        action.

A side feature is that ```views.py``` becomes the entry point for any
view regarding models and has to import other files (model-views) that
will implement the actual indexing or detail functions.

As a result, this commit also introduces ```users.py```, ```vms.py```
and ```projects.py```. The actions are instantiated in these files.

Fixes #28, although it does not follow the proposed solution to this
problem (JSON) and uses classes instead.