admin: Update admin logout
In order to logout properly without using the astakos' urls, we will use
astakosclient.

More specifically, we will use the 'ui_url' property of astakosclient
that holds the url for the UI of Astakos and we will append to the end
of this url the '/logout' string. The user that requests to logout will
therefore get redirected there.

Fix apyrgio/synnefo#117