admin: Add RBAC
The definition of which groups can author actions is kept in the conf
file of admin and has the following structure:

ADMIN_RBAC = {
    'user': {
        'activate': ['admin', 'helpdesk'],
        .
        .
        .
    },
    'vm': {...},
    .
    .
    .
}

That is, the main entry point is ADMIN_RBAC, which is a dictionary. The
first level of this dictionary is the target category of an action, e.g.
user, vm, volume etc. The second level is the action, which points to a
list of groups that can author it.

Fix apyrgio/synnefo#149