admin: Add IP log in item associations
The IPAddressLog model is associated with the following models:

* VirtualMachine
* Network
* IPAdress

Given that VMs have a 1:1 association with users, this means that the IP
log is also associated with users.

Using this information, we can show a customised IP log in the details
view of each of the above models:

* Users: we will show the IPs that they have allocated.
* VMs: we will show the IPs that have been attached to them.
* Networks: we will show the IPs that have belonged at some point to a
  network.
* IPs: We will show to which servers and networks an IP has been
  attached.

Every log entry, regardless of where it's shown, will have the
following info:

* IP address,
* VM
* Network
* User
* Active indication (if the IP is still has not been detached)
* Date of attachment to the server
* Date of release.

All log entries will be shown in chronological order and will point to
the respective associated items.

Fix apyrgio/synnefo#133.