Bug with input for users
The functionality of this input is both in controller and view.
For each group there is an input-emails. If you try to add something in
it, the users of every input view are updated.

Users come from the usersExtended enumerable, that is a prop of
GroupController.

addUser function changes the value of usersExtended.

The bug seems to be fixed with the init function that I have in
comments, but still seems to be abnormal behavor.